Removing libstlport_shared.so and updating Makefile to pull libstlport_shared.so from the ${ANDROID_NDK_ROOT} path.
The copy of libstlport_shared.so is different with the latest NDK release (revision 7)
and the binary .so files are not compatible with the older version of libstlport_shared.so.
Now we should package the libstlport_shared.so available from the build environment.

Note the changes to the Makefile require the ANDROID_NDK_ROOT environment variable be defined.