Dropping setUseHMACPageProtection in favor of SQLiteDatabaseHook.
An optional SQLiteDatabaseHook interface is introduced.  This allows
for abitrary execution both before and after the PRAGMA key is set,
but prior to the setLocale allowing users to configure their
instance accordingly.  The preKey operation will likely be used by
the cipher_default_use_hmac PRAGMA.  The postKey callback allows
for operations such as setting the cipher PRAGMA.