add build metadata to jar file to aid reproducible builds
This moves jar creation to `ant release`.

To provide build security, it should be possible for anyone to run the
build process and end up with the same result as Zetetic.  To aid that
process, this adds metadata to the release jar that specifies which
versions of Java, Android SDK, and Android NDK were used to build the
official release.