use faketime, if available, to set the build time based on the git commit
If the `faketime` utility is available, use it to fix the time of the build
based on the time of the current git commit.  This makes the build process
reproducible since anyone who builds it from git will end up with a build
that has the same exact timestamps.

You can get faketime from various package systems:
 apt-get install faketime
 brew install faketime
 port install libfaketime