use `git describe` for release version and fix type in zipball name
sqlipher --> sqlcipher

`git describe` will give only the name of the tag if the current commit
is tagged.  Otherwise, it will get the most recent tag, how many commits
since that tag, and the commit hash id.   This marks non-release builds
clearly and enforces that the release must be made from the proper tagged
git commit.