find path to set ndk.dir, if ndk.dir is not set already
One standard way of setting ndk.dir is to set it in a .properties file. If
that has not been done, then this ant code looks for the env var
ANDROID_NDK and the path where ndk-build is installed.

ndk.dir is not currently used by the build system, except to write out info
to the manifest about the build environment.