Support optional corrupt database handler (ref: sqlcipher/android-database-sqlcipher#169).
Add additional safeguards to public SQLiteDatabase.markTableSyncable() functions in case the database is not opened.

NOTE: The DefaultDatabaseErrorHandler follows the existing implementation from SQLiteDatabase.onCorruption()
which is to close the database object and then delete it. This is different from the AOSP version
of DefaultDatabaseErrorHandler.java which also attempts to delete all attached database files.
FUTURE/TBD I am not sure if we want to do this or not.