Limit memory usage, improve cursor read throughput.
Migrated over the original implementation of CursorWindow#allocate so
that it respects the maximum memory limit.

Also, increased ROW_SLOT_CHUNK_NUM_ROWS to 256, improving allocation
speed and drastically increasing the throughput of cursor reads. That
value determines how many row_slot_t's are allocated, which is just a
struct wrapping a uint32_t, so increasing the number allocated at once
shouldn't be a huge deal from a memory usage perspective. The value 256
was chosen after testing with a variety of devices. It appeared to be
the size where you had diminishing returns on cursor filling
performance.