Allow for improved cursor window loading heuristic
The underlying SQLiteCursor implementation will request the
CursorWindow to fill based on a computed heuristic which attempts to
load 1/3 of the rows preceeding to the requested index. This shows
a marked improvment in data loading. The new loading heuristic will be
the default behavior. This can be disabled by invoking the following on
the underlying SQLiteCursor reference:

SQLiteCursor cursor = ...
cursor.setFillWindowForwardOnly(true);

The backing memory used to store the content of the CursorWindow can
be modified from the default behavior setting the initial, growth padding,
and maximum memory to be allocated during the mapping process into the
CursorWindow. This can be accomplished via the following:

long initial = 1024 * 1024;
long growthPadding = 1024 * 1024 * 2;
long max = 1024 * 1024 * 4;
CursorWindowAllocation allocation =
  new CustomCursorWindowAllocation(initial, growthPadding, max);
CursorWindow.setCursorWindowAllocation(allocation);